requirements.txt
# --- Core LLM & Framework ---
langchain>=0.2.11
langchain-core>=0.2.23
langchain-community>=0.2.10
langchain-openai>=0.1.17
langchain-google-genai>=1.0.8
langgraph>=0.1.4

# --- Search & Tools ---
tavily-python>=0.3.5
google-generativeai>=0.7.2

# --- Web Scraping & Document Loading ---
beautifulsoup4>=4.12.3
requests>=2.32.3
lxml>=5.2.2
html5lib>=1.1

# --- Async & Performance ---
aiohttp>=3.9.5
orjson>=3.10.6
tenacity>=8.5.0

# --- Data & Validation ---
pandas>=2.2.2
pydantic>=2.8.2
python-dotenv>=1.0.1

# --- Observability & Utilities ---
tiktoken>=0.7.0
rich>=13.7.1